+++ api/v1alpha1/silence_types.go
@@ -26,10 +26,14 @@ type SilenceSpec struct {
 	Matchers   []Matcher   `json:"matchers"`
 
 	// Owner is GitHub username of a person who created and/or owns the silence.
-	Owner *string `json:"owner,omitempty"`
+	Owner string `json:"owner,omitempty"`
 
 	// PostmortemURL is a link to a document describing the problem.
+	// Deprecated: Use IssueURL instead.
 	PostmortemURL *string `json:"postmortem_url,omitempty"`
+
+	// IssueURL is a link to a GitHub issue describing the problem.
+	IssueURL string `json:"issue_url,omitempty"`
 }
 
 type TargetTag struct {